fw-5510, display song introduction + intro translation
- move convertJsonToComponent to common function
- display song introductions + introduction translations
